+++ packages/serverless-components/nextjs-component/types.d.ts
@@ -1,4 +1,4 @@
-import { PublicDirectoryCache } from "@sls-next/s3-static-assets/src/lib/getPublicAssetCacheControl";
+import { PublicDirectoryCache } from "@contentco/s3-static-assets/src/lib/getPublicAssetCacheControl";
 
 export type ServerlessComponentInputs = {
   build?: BuildOptions | boolean;