+++ packages/libs/s3-static-assets/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.3](https://github.com/danielcondemarin/serverless-next.js/compare/@contentco/s3-static-assets@1.3.2...@contentco/s3-static-assets@1.3.3) (2020-09-05)
+
+**Note:** Version bump only for package @contentco/s3-static-assets
+
 ## [1.3.2](https://github.com/danielcondemarin/serverless-next.js/compare/@contentco/s3-static-assets@1.3.1...@contentco/s3-static-assets@1.3.2) (2020-09-05)
 
 **Note:** Version bump only for package @contentco/s3-static-assets