+++ packages/serverless-components/domain/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.5](https://github.com/himynameistimli/serverless-next.js/compare/@contentco/domain@1.0.4...@contentco/domain@1.0.5) (2020-09-05)
+
+**Note:** Version bump only for package @contentco/domain
+
 ## [1.0.4](https://github.com/himynameistimli/serverless-next.js/compare/@contentco/domain@1.0.3...@contentco/domain@1.0.4) (2020-09-05)
 
 **Note:** Version bump only for package @contentco/domain